DRY up resources too
We're repeating ourselves a lot here, so I've added a `Resource` class that
all the resources inherit from. Each resource class has a
`SUPPORTED_ENDPOINTS` class, and if an unupported method gets used,
it raises a `NotSupportedError`
