+++ lib/breathe.rb
@@ -3,6 +3,7 @@ require "sawyer"
 require "breathe/version"
 require "breathe/client"
 require "breathe/response"
+require "breathe/resource"
 
 require "breathe/absences"
 require "breathe/sicknesses"
@@ -13,5 +14,6 @@ module Breathe
   class Error < StandardError; end
   class AuthenticationError < StandardError; end
   class UnknownError < StandardError; end
+  class NotSupportedError < StandardError; end
   # Your code goes here...
 end
