+++ lib/breathe/absences.rb
@@ -1,17 +1,8 @@
 module Breathe
-  class Absences
-    attr_reader :client
-
-    def initialize(client)
-      @client = client
-    end
-
-    def list(args = {})
-      client.response(
-        method: :get,
-        path: "absences",
-        args: args
-      )
-    end
+  class Absences < Resource
+    RESOURCE_NAME = "absences"
+    SUPPORTED_ENDPOINTS = [
+      :list,
+    ]
   end
 end
