+++ lib/breathe/employees.rb
@@ -1,24 +1,9 @@
 module Breathe
-  class Employees
-    attr_reader :client
-
-    def initialize(client)
-      @client = client
-    end
-
-    def list(args = {})
-      client.response(
-        method: :get,
-        path: "employees",
-        args: args
-      )
-    end
-
-    def get(id)
-      client.response(
-        method: :get,
-        path: "employees/#{id.to_i}"
-      )
-    end
+  class Employees < Resource
+    RESOURCE_NAME = "employees"
+    SUPPORTED_ENDPOINTS = [
+      :list,
+      :get,
+    ]
   end
 end
