+++ lib/breathe/sicknesses.rb
@@ -1,17 +1,8 @@
 module Breathe
-  class Sicknesses
-    attr_reader :client
-
-    def initialize(client)
-      @client = client
-    end
-
-    def list(args = {})
-      client.response(
-        method: :get,
-        path: "sicknesses",
-        args: args
-      )
-    end
+  class Sicknesses < Resource
+    RESOURCE_NAME = "sicknesses"
+    SUPPORTED_ENDPOINTS = [
+      :list,
+    ]
   end
 end
